start to header

+++ be/models/movie.js
@@ -4,6 +4,7 @@ const PATH = require('path') // 路径
 
 // 创建的Model模型 （collection）
 //在本地数据库中创建一个document
+// movies，不加s，mongoose会自动加s，为了不懵逼，就加上
 var MovieModel = mongoose.model('movies', new mongoose.Schema({
     //定义集合中存储的数据名，数据格式
     movieName: String,
@@ -18,18 +19,23 @@ var MovieModel = mongoose.model('movies', new mongoose.Schema({
 
 // 返回列表数据
 const list = async ({ pageNo = 1, pageSize = 5, search = '' }) => {
-    let re = new RegExp(search, 'i')
-    let _query = search ?  { movieName: re } : {}// 查询的约定条件
-    // console.log(_query, 111)
+    let reg = new RegExp(search, 'g')
+    let _query = {      // 查询的约定条件
+        $or:[
+            { movieName: reg },   
+            { directorName: reg },   
+            { startName: reg },   
+            { movieType: reg }
+        ]
+    }
     // limit // 取几条
     // skip // 从哪里开始
     let _all_items = await listall(_query)
 
-
     return MovieModel.find(_query)
     .sort({createTime: -1})
     .skip((pageNo - 1) * pageSize)// 从哪一页开始
-    .limit(~~pageSize)// 截取多少
+    .limit(~~pageSize)// 截取多少，从这一页开始取多少个数据
     .then((results) => {
         return { 
             items: results, 
@@ -37,7 +43,8 @@ const list = async ({ pageNo = 1, pageSize = 5, search = '' }) => {
                 pageNo, // 当前页
                 pageSize, // 一页数量
                 total: _all_items.length, // 总数
-                totalPage: Math.ceil(_all_items.length / pageSize) // 总页数
+                totalPage: Math.ceil(_all_items.length / pageSize), // 总页数
+                search  // 搜索关键字
             }
         }
     }).catch((err) => {
@@ -47,9 +54,8 @@ const list = async ({ pageNo = 1, pageSize = 5, search = '' }) => {
 
 
 // 返回列表所有数据
-const listall = (query) => {
+const listall = (_query = {}) => {
     //_query，查找数据的规则，eg：年龄大于10……，为空则全部返回
-    let _query= query || {}
     //Model.fin()查询数据库
     //mongodb的方法：sort({createTime:-1})数据倒序返回
     return MovieModel.find(_query).sort({createTime:-1})
@@ -83,7 +89,7 @@ const save = (body) => {
 }
 
 // 返回一个数据
-const listone = ({ id }) => {
+const listone = ({ id }) => { 
     return MovieModel.findById(id).
         then((results) => {     //返回数据库的数据
             return results
@@ -105,8 +111,9 @@ const update =async (body)=>{
     //更新后把原来的图片先删掉
     let {id} = body  
     let _row = await listone({id})
-    fs.removeSync(PATH.resolve(__dirname,'../public'+_row.movieLogo))
-
+    if (_row.companyLogo &&_row.movieLogo != default_logo) {
+        fs.removeSync(PATH.resolve(__dirname,'../public'+_row.movieLogo))    
+    }
     return MovieModel.updateOne({ _id: body.id }, { ...body }).then((results) => {
         return results
     }).catch((err) => {
@@ -115,11 +122,18 @@ const update =async (body)=>{
 }   
 
 //删除一条数据
-const remove = async ({ id })=>{
+const remove = async ({ id, pageNo, pageSize })=>{
     let _row = await listone({ id })    //用listone从数据库中获得一条数据的信息
-    return MovieModel.deleteOne({ _id:id }).then((results)=>{
+    return MovieModel.deleteOne({ _id:id }).then( async (results)=>{
+
+        //  获取最新的数量,删除的过程中有新增数据的产生
+        let _all_items = await listall()
+        
+        results.removeId = id
+        results.isBack = (pageNo-1) * pageSize >= _all_items.length  
+
         results.removeId = id   //这个id是返回给前端用的
-        if (_row.movieLogo != default_logo) {
+        if (_row.companyLogo &&_row.movieLogo != default_logo) {
             fs.removeSync(PATH.resolve(__dirname,'../public'+_row.movieLogo))    
         }
         return results
